+++ R/ci_wald.R
@@ -51,8 +51,14 @@ ci_wald <- function(model, ci = .95, dof = NULL, effects = c("fixed", "random", 
       )
     }
 
+    if (.is_empty_object(stderror)) {
+      return(NULL)
+    }
+
     # filter non-matching parameters
-    stderror <- stderror[1:nrow(params), ]
+    if (nrow(stderror) != nrow(params)) {
+      stderror <- merge(stderror, params, sort = FALSE)
+    }
     se <- stderror$SE
   }