+++ inst/doc/efa_cfa.R
@@ -74,3 +74,23 @@ big6 <- lavaan::cfa(structure_big6, data = test)
 
 performance::compare_performance(big5, big6)
 
+## ----message=FALSE, warning=FALSE---------------------------------------------
+set.seed(1234)
+X <- rnorm(100)
+M <- 0.5 * X + rnorm(100)
+Y <- 0.7 * M + rnorm(100)
+Data <- data.frame(X = X, Y = Y, M = M)
+
+model <- ' # direct effect
+             Y ~ c*X
+           # mediator
+             M ~ a*X
+             Y ~ b*M
+           # indirect effect (a*b)
+             ab := a*b
+           # total effect
+             total := c + (a*b)
+         '
+fit <- sem(model, data = Data, test = "Satorra-Bentler")
+model_parameters(fit)
+